Separated ANE code from Android YouTube code Instantiating YouTubePlayer again for case where fragment is disposed and recreated
