+++ android/src/meez/nativeExtensions/androidyoutube/AndroidYoutubeContext.java
@@ -4,98 +4,32 @@ import java.util.*;
 import java.util.concurrent.atomic.AtomicInteger;
 
 import android.annotation.TargetApi;
-import android.app.Dialog;
 import android.content.pm.ApplicationInfo;
 import android.content.pm.PackageManager;
-import android.graphics.Color;
-import android.graphics.drawable.ColorDrawable;
 import android.os.Build;
-import android.os.Bundle;
-import android.os.Handler;
 import android.util.Log;
-import android.view.*;
-import android.widget.FrameLayout;
 import com.adobe.fre.FREContext;
 import com.adobe.fre.FREFunction;
 import com.adobe.fre.FREObject;
-import com.google.android.youtube.player.YouTubeInitializationResult;
 import com.google.android.youtube.player.YouTubePlayer;
-import com.google.android.youtube.player.YouTubePlayerFragment;
 import org.json.JSONObject;
 
 @TargetApi(Build.VERSION_CODES.HONEYCOMB)
-public class AndroidYoutubeContext extends FREContext implements YouTubePlayer.OnInitializedListener, YouTubePlayer.PlayerStateChangeListener, YouTubePlayer.PlaybackEventListener
+public class AndroidYoutubeContext extends FREContext implements YouTubePlayerAdapter.PlayerEventCallback
 {
     // Definitions
-
-    /** ANE State */
-    private enum State
-    {
-        INITIALIZING,
-        READY,
-        DISPOSING,
-        DISPOSED
-    }
-
-    /** Player States */
-    public static final String PLAYER_STATE_PLAYING="playing";
-    public static final String PLAYER_STATE_UNSTARTED="unstarted";
-    public static final String PLAYER_STATE_ENDED="ended";
-    public static final String PLAYER_STATE_PAUSED="paused";
-    public static final String PLAYER_STATE_BUFFERING="buffering";
-    public static final String PLAYER_STATE_CUED="cued";
-    public static final String PLAYER_STATE_UNKNOWN="unknown";
-
-    /** Player Errors */
-    public static final String PLAYER_ERROR_INVALID="invalid";
-    public static final String PLAYER_ERROR_HTML5="html5";
-    public static final String PLAYER_ERROR_NOT_FOUND="not found";
-    public static final String PLAYER_ERROR_EMBED="cannot embed";
-    public static final String PLAYER_ERROR_NETWORK="network error";
-    public static final String PLAYER_ERROR_TOO_SMALL="player view too small";
-    public static final String PLAYER_ERROR_OVERLAID="unauthorized overlay";
-
-    /** YouTube Player Fragment */
-    private YouTubePlayerFragment playerFragment;
-
-    /** YouTube Player */
-    private YouTubePlayer player;
-
-    /** Dialog Box to hold video player view */
-    private Dialog dialog;
-
-    /** Layout Parameters */
-    private WindowManager.LayoutParams layoutParams;
-
-    /** Video Time Timer */
-    private VideoTimer videoTimer;
-
-    /** Video Container */
-    private FrameLayout videoContainer;
 
     /** Reason Video player is unsupported */
     private String unsupportedReason;
 
-    /** Current State */
-    private State state;
-
-    /** Developer Key */
-    private String devKey;
-
-    // temp vars to hold onto until player is created
-
-    /** URL/VideoId to load in Player */
-    private String videoToPlay;
-
-    /** Video Start time */
-    private int startTime = 0;
+    /** Adapter */
+    private YouTubePlayerAdapter adapter;
 
     /** Create a new AndroidYoutubeContext */
     public AndroidYoutubeContext()
     {
         Log.d(Extension.TAG, "AndroidYouTubeContext()");
 
-        changeState(State.INITIALIZING);
         this.unsupportedReason="";
     }
 
@@ -105,32 +39,7 @@ public class AndroidYoutubeContext extends FREContext implements YouTubePlayer.O
     {
         Log.d(Extension.TAG, "AndroidYouTubeContext.dispose()");
 
-        changeState(State.DISPOSING);
-
-        stopUpdateTimer();
-
-        getRootContainer().removeView(videoContainer);
-        this.videoContainer.removeAllViews();
-        this.dialog.dismiss();
-        this.playerFragment.setRetainInstance(false);
-
-        try
-        {
-            // Removing fragment will call onDestroyView() on the fragment, which will release the YouTubePlayer
-            // see https://developers.google.com/youtube/android/player/reference/com/google/android/youtube/player/YouTubePlayerFragment#Overview
-            getActivity().getFragmentManager()
-                    .beginTransaction()
-                    .remove(this.playerFragment)
-                    .commit();
-        }
-        catch (Throwable t)
-        {
-            Log.w(Extension.TAG, "Error removing player fragment", t);
-        }
-        finally
-        {
-            changeState(State.DISPOSED);
-        }
+        YouTubePlayerAdapter.releaseInstance();
     }
 
     /** Registers AS function name to Java Function Class */
@@ -151,185 +60,69 @@ public class AndroidYoutubeContext extends FREContext implements YouTubePlayer.O
         return functionMap;
     }
 
-    /**
-     * Initialize the video player
-     * @param devKey Developer key from google/youtube
-     */
-    public void initVideo(String devKey)
-    {
-        Log.d(Extension.TAG, "AndroidYouTubeContext.initVideo()");
-
-        // Should only be called 1x during init state
-        assertState(State.INITIALIZING);
-
-        this.devKey = devKey;
-
-        // Create views
-        this.videoContainer = createVideoContainer();
-        this.videoContainer.setId(AndroidYoutubeContext.generateViewId());
-        this.layoutParams = createLayoutParams();
-        this.dialog = createDialog();
-        this.playerFragment = createPlayerFragment();
-
-        //Hack - videoContainer is only added to root view long enough to attach fragment (fragment must be attached to view in Android 'display list').
-        // In Fragment's onViewCreated() videoContainer is removed from root container and set as view of Dialog
-        this.videoContainer.setVisibility(View.INVISIBLE);
-        getRootContainer().addView(this.videoContainer);
-
-        // Add YouTube fragment to video container
-        getActivity()
-                .getFragmentManager()
-                .beginTransaction()
-                .add(this.videoContainer.getId(), this.playerFragment)
-                .commit();
-    }
-
-    /** On Fragment View Created */
-    public void onFragmentViewCreated()
-    {
-        // Called from Youtube Fragment
-
-        if (!checkState(State.INITIALIZING))
-            return;
-
-        // Remove from root and add to dialog
-        getRootContainer().removeView(videoContainer);
-
-        // hide video offscreen until explicitly positioned by Actionscript call
-        setVideoFrame(-5000, -5000, 600, 400);
-
-        this.dialog.setContentView(videoContainer);
-        this.dialog.show();
-
-        this.videoContainer.setVisibility(View.VISIBLE);
-
-        changeState(State.READY);
-
-        // Initialize the video player fragment
-        this.playerFragment.initialize(this.devKey, this);
-    }
-
-    /** On Fragment Save Instance */
-    public void onFragmentSaveInstance(boolean saveState)
-    {
-        // Called from Youtube Fragment
-
-        setVideoSaveEnabled(saveState);
-    }
-
-    // Views
-
-    /** Get Root Container ViewGroup */
-    public ViewGroup getRootContainer()
-    {
-        return (ViewGroup)((ViewGroup)getActivity().findViewById(android.R.id.content)).getChildAt(0);
-    }
-
-    /** Create Video Container Layout */
-    public FrameLayout createVideoContainer()
-    {
-        FrameLayout.LayoutParams params = new FrameLayout.LayoutParams(ViewGroup.LayoutParams.MATCH_PARENT, ViewGroup.LayoutParams.MATCH_PARENT);
-        FrameLayout layout = new FrameLayout(getActivity());
-        layout.setLayoutParams(params);
-        return  layout;
-    }
-
-    /** Create YouTube Player Fragment */
-    public YouTubePlayerFragment createPlayerFragment()
-    {
-        AndroidYoutubeContext.CustomPlayerFragment fragment = new AndroidYoutubeContext.CustomPlayerFragment();
-        fragment.setRetainInstance(true);
-        return fragment;
-    }
-
-    /** Create Video layout params */
-    public WindowManager.LayoutParams createLayoutParams()
-    {
-        WindowManager.LayoutParams params = new WindowManager.LayoutParams();
-        params.gravity = Gravity.TOP | Gravity.LEFT;
-        params.dimAmount = 0.0f;
-        params.flags = WindowManager.LayoutParams.FLAG_NOT_FOCUSABLE |
-                WindowManager.LayoutParams.FLAG_LAYOUT_NO_LIMITS |
-                WindowManager.LayoutParams.FLAG_NOT_TOUCH_MODAL |
-                WindowManager.LayoutParams.FLAG_ALT_FOCUSABLE_IM;
-        return params;
-    }
-
-    /** Create Video Dialog  */
-    public Dialog createDialog()
-    {
-        Dialog dialog = new Dialog(getActivity());
-        dialog.setCancelable(false);
-        dialog.requestWindowFeature(Window.FEATURE_NO_TITLE);
-        dialog.getWindow().clearFlags(WindowManager.LayoutParams.FLAG_DIM_BEHIND);
-        dialog.getWindow().setBackgroundDrawable(new ColorDrawable(Color.TRANSPARENT));
-        return dialog;
-    }
+    // Create PlayerActions
 
     /** Play video by ID (YouTube video id) */
-    public void playVideoById(String videoId, double startTime)
+    public YouTubePlayerAdapter.PlayerAction createPlayByIdAction(final String videoId, final double startTime)
     {
-        Log.d(Extension.TAG, "AndroidYouTubeContext.playVideoById("+videoId+")");
+        return new YouTubePlayerAdapter.PlayerAction() {
 
-        // seconds to ms
-        int startTimeMS = (int)(1000 * startTime);
-        if (playerIsReady())
-        {
-            this.player.loadVideo(videoId, startTimeMS);
-        }
-        else
-        {
-            // store to play when player is ready
-            this.videoToPlay = videoId;
-            this.startTime = startTimeMS;
-        }
+            @Override
+            public void run(YouTubePlayer player) throws Exception
+            {
+                Log.d(Extension.TAG, "AndroidYouTubeContext.playVideoById("+videoId+")");
+
+                player.loadVideo(videoId,(int)(1000 * startTime));
+            }
+        };
     }
 
     /** Play Video */
-    public void playVideo()
+    public YouTubePlayerAdapter.PlayerAction createPlayVideoAction()
     {
-        if (!playerIsReady())
-            return;
-        this.player.play();
+        return new YouTubePlayerAdapter.PlayerAction()
+        {
+            @Override
+            public void run(YouTubePlayer player) throws Exception
+            {
+                player.play();
+            }
+        };
     }
 
     /** Pause Video */
-    public void pauseVideo()
+    public YouTubePlayerAdapter.PlayerAction createPauseAction()
     {
-        if (!playerIsReady())
-            return;
-        this.player.pause();
+        return new YouTubePlayerAdapter.PlayerAction() {
+
+            @Override
+            public void run(YouTubePlayer player) throws Exception
+            {
+                player.pause();
+            }
+        };
     }
 
     /** Seek in Video */
-    public void doSeek(double timeInSeconds)
+    public YouTubePlayerAdapter.PlayerAction createSeekAction(final double time) throws Exception
     {
-        if (!playerIsReady())
-            return;
-
-        int timeInMilliseconds = (int)(1000 * timeInSeconds);
-        int dur = this.player.getDurationMillis();
-
-        // prevent seeking past video end
-        if (dur>0 && timeInMilliseconds>=dur)
+        return new YouTubePlayerAdapter.PlayerAction()
         {
-            // go back 1 second from video end
-            timeInMilliseconds = Math.max(0, dur-1000);
-        }
-        this.player.seekToMillis(timeInMilliseconds);
-    }
+            @Override
+            public void run(YouTubePlayer player) throws Exception
+            {
+                int dur = player.getDurationMillis();
+                int timeInMilliseconds = (int)(1000 * time);
 
-    /** Set video frame */
-    public void setVideoFrame(int x, int y, int width, int height)
-    {
-        Log.d(Extension.TAG, String.format("AndroidYouTubeContext.setVideoFrame(%d, %d, %d, %d)", x, y, width, height));
-
-        this.layoutParams.width = width;
-        this.layoutParams.height = height;
-        this.layoutParams.x = x;
-        this.layoutParams.y = y;
-
-        this.dialog.getWindow().setAttributes(this.layoutParams);
+                // prevent seeking past video end
+                if (dur>0 && timeInMilliseconds>=dur)
+                {
+                    // go back 1 second from video end
+                    timeInMilliseconds = Math.max(0, dur-1000);
+                }
+                player.seekToMillis(timeInMilliseconds);
+            }
+        };
     }
 
     /** Get the reason YouTube video is unsupported */
@@ -350,27 +143,29 @@ public class AndroidYoutubeContext extends FREContext implements YouTubePlayer.O
 
         if (action.equals("play"))
         {
-            playVideo();
-        }
+            this.adapter.execute(createPlayVideoAction());
 
-        else if(action.equals("playById"))
+        }else if(action.equals("playById"))
         {
-            playVideoById(msgObj.getString("videoId"), msgObj.getDouble("startTime"));
+            String videoId=msgObj.getString("videoId");
+            double startTime = msgObj.getDouble("startTime");
+
+            this.adapter.execute(createPlayByIdAction(videoId,startTime));
         }
 
         else if(action.equals("stop"))
         {
-            pauseVideo();
+            this.adapter.execute(createPauseAction());
         }
 
         else if(action.equals("pause"))
         {
-            pauseVideo();
+            this.adapter.execute(createPauseAction());
         }
 
         else if(action.equals("seek"))
         {
-            doSeek(msgObj.getDouble("time"));
+            this.adapter.execute(createSeekAction(msgObj.getDouble("time")));
         }
 
         else if(action.equals("dispose"))
@@ -385,270 +180,7 @@ public class AndroidYoutubeContext extends FREContext implements YouTubePlayer.O
         }
     }
 
-    // YouTubePlayer.OnInitializedListener Implementation
-
-    @Override
-    public void onInitializationFailure(YouTubePlayer.Provider provider, YouTubeInitializationResult errorReason)
-    {
-        Log.w(Extension.TAG, "AndroidYouTubeContext.onInitializaionFailure(). Could not Initialize player. Reason ("+errorReason.toString()+")");
-
-        dispatchEventWithReason("videoError", "Could not initialize video player");
-    }
-
-    @Override
-    public void onInitializationSuccess(YouTubePlayer.Provider provider, YouTubePlayer player, boolean wasRestored)
-    {
-        Log.d(Extension.TAG, "AndroidYouTubeContext.onInitSuccess (player="+player+", wasRestored="+wasRestored+")");
-
-        if (!checkState(State.READY))
-            return;
-
-        this.player=player;
-        this.player.setPlayerStateChangeListener(this);
-        this.player.setPlaybackEventListener(this);
-
-        if (!wasRestored)
-        {
-            // no video player controls
-            this.player.setPlayerStyle(YouTubePlayer.PlayerStyle.CHROMELESS);
-
-            if (this.videoToPlay!=null)
-            {
-                this.player.loadVideo(this.videoToPlay, this.startTime);
-            }
-            this.videoToPlay = null;
-            this.startTime = 0;
-        }
-
-        sendData("playerReady");
-    }
-
-    //  YouTubePlayer.PlayerStateChangeListener Implementation
-
-    @Override
-    public void onAdStarted()
-    {
-        Log.d(Extension.TAG, "Ad Started");
-        sendData("Ad started");
-    }
-
-    @Override
-    public void onError(YouTubePlayer.ErrorReason errorReason)
-    {
-        Log.w(Extension.TAG, "AndroidYouTubeContext.onError() error: ("+errorReason.toString()+")");
-
-        String msg;
-        int code;
-
-        // 1st two are not counted as errors, but Actionscript player needs to be notified
-        if (errorReason==YouTubePlayer.ErrorReason.UNAUTHORIZED_OVERLAY)
-        {
-            msg=PLAYER_ERROR_OVERLAID;
-            code=0;
-        }
-        else if (errorReason==YouTubePlayer.ErrorReason.PLAYER_VIEW_TOO_SMALL)
-        {
-            msg=PLAYER_ERROR_TOO_SMALL;
-            code=0;
-        }
-
-        else if (errorReason==YouTubePlayer.ErrorReason.NOT_PLAYABLE ||
-                 errorReason==YouTubePlayer.ErrorReason.AUTOPLAY_DISABLED ||
-                 errorReason==YouTubePlayer.ErrorReason.INTERNAL_ERROR)
-        {
-            msg=PLAYER_ERROR_INVALID;
-            code=2;
-        }
-
-        else if (errorReason==YouTubePlayer.ErrorReason.EMPTY_PLAYLIST)
-        {
-            msg=PLAYER_ERROR_NOT_FOUND;
-            code=100;
-        }
-
-        else if (errorReason==YouTubePlayer.ErrorReason.USER_DECLINED_RESTRICTED_CONTENT ||
-                 errorReason==YouTubePlayer.ErrorReason.USER_DECLINED_HIGH_BANDWIDTH)
-        {
-            msg=PLAYER_ERROR_EMBED;
-            code=101;
-        }
-        else if (errorReason==YouTubePlayer.ErrorReason.NETWORK_ERROR ||
-                 errorReason==YouTubePlayer.ErrorReason.UNEXPECTED_SERVICE_DISCONNECTION)
-        {
-            msg=PLAYER_ERROR_NETWORK;
-            code=500; // Meez API Error Code
-        }
-        // Unknown reason
-        else
-        {
-            msg=errorReason.toString();
-            code=500; // Meez API Error Code
-        }
-        sendError(code, msg);
-    }
-
-    @Override
-    public void onLoaded(String arg0)
-    {
-        Log.d(Extension.TAG, "AndroidYouTubeContext.onLoaded("+arg0+")");
-
-        sendState(PLAYER_STATE_UNSTARTED);
-    }
-
-    @Override
-    public void onLoading()
-    {
-        Log.d(Extension.TAG, "AndroidYouTubeContext.onLoading");
-
-        sendState(PLAYER_STATE_BUFFERING);
-    }
-
-    @Override
-    public void onVideoEnded()
-    {
-        Log.d(Extension.TAG, "AndroidYouTubeContext.onVideoEnded");
-
-        sendState(PLAYER_STATE_ENDED);
-    }
-
-    @Override
-    public void onVideoStarted()
-    {
-        Log.d(Extension.TAG, "AndroidYouTubeContext.onVideoStarted");
-        // Nothing to send
-    }
-
-    // YouTubePlayer.PlaybackEventListener Implementation
-
-    @Override
-    public void onBuffering(boolean isBuffering)
-    {
-        if (isBuffering)
-        {
-            sendState(PLAYER_STATE_BUFFERING);
-        }
-    }
-
-    @Override
-    public void onPaused()
-    {
-        Log.d(Extension.TAG, "AndroidYouTubeContext.onPaused");
-
-        stopUpdateTimer();
-        sendState(PLAYER_STATE_PAUSED);
-    }
-
-    @Override
-    public void onPlaying()
-    {
-        Log.d(Extension.TAG, "AndroidYouTubeContext.onPlaying");
-
-        startUpdateTimer();
-        sendState(PLAYER_STATE_PLAYING);
-    }
-
-    @Override
-    public void onSeekTo(int newPositionMillis)
-    {
-        Log.d(Extension.TAG, "AndroidYouTubeContext.onSeekTo("+Integer.toString(newPositionMillis)+")");
-    }
-
-    @Override
-    public void onStopped()
-    {
-        Log.d(Extension.TAG, "AndroidYouTubeContext.onStopped");
-
-        stopUpdateTimer();
-    }
-
-    // Implementation
-
-    /** Callable on View Interface */
-    public interface CallableOnView
-    {
-        void call(View view);
-    }
-
-    /** Set Save Enabled on Video view and its children */
-    protected void setVideoSaveEnabled(final boolean value)
-    {
-        recursiveCallOnView(this.videoContainer, new CallableOnView(){
-            @Override public void call(View view) {
-                view.setSaveEnabled(value);
-            }
-        });
-    }
-
-    /** Recursively call action on View */
-    protected void recursiveCallOnView(View view, CallableOnView callableOnView)
-    {
-        if (view==null)
-            return;
-
-        if (view instanceof ViewGroup)
-        {
-            for (int i=0; i < ((ViewGroup) view).getChildCount(); i++)
-            {
-                recursiveCallOnView(((ViewGroup) view).getChildAt(i), callableOnView);
-            }
-        }
-        callableOnView.call(view);
-    }
-
-    /** Start the update timer */
-    private void startUpdateTimer()
-    {
-        if (this.videoTimer==null)
-        {
-            this.videoTimer = new VideoTimer(1000L);
-        }
-        this.videoTimer.start();
-    }
-
-    /** Stop the update timer */
-    private void stopUpdateTimer()
-    {
-        if (this.videoTimer!=null)
-        {
-            this.videoTimer.stop();
-        }
-    }
-
-    /** Player is ready */
-    private boolean playerIsReady()
-    {
-        return this.player!=null;
-    }
-
-    /** Change state */
-    private State changeState(State newState)
-    {
-        State oldState = this.state;
-        this.state = newState;
-        return oldState;
-    }
-
-    /** Check the current state */
-    private boolean checkState(State expState)
-    {
-        boolean correct=this.state.equals(expState);
-        if (!correct)
-        {
-            Log.w(Extension.TAG, "Check state failed. Expected ("+expState+"). Actual ("+this.state+")");
-        }
-        return correct;
-    }
-
-    /** Assert a State */
-    private void assertState(State s)
-    {
-        if (!checkState(s))
-        {
-            throw new IllegalStateException("Expected state("+s+"). Actual state ("+this.state+")");
-        }
-    }
-
-    // Event dispatching
+    // YouTubePlayerAdapter.PlayerEventCallback impl
 
     /** Send JSON encoded Message (to Actionscript) */
     public void sendMessage(JSONObject msg)
@@ -658,7 +190,7 @@ public class AndroidYoutubeContext extends FREContext implements YouTubePlayer.O
     }
 
     /** Send generic data to Actionscript */
-    private void sendData(String data)
+    public void sendData(String data)
     {
         //sendMessage({type:"data", data:data});
         try
@@ -675,27 +207,27 @@ public class AndroidYoutubeContext extends FREContext implements YouTubePlayer.O
     }
 
     /** Send Time data to Actionscript */
-    private void sendTimeData() throws Exception
+    public void sendTimeData(int curSecs, int totalSecs)
     {
-        // if not playing, send no data
-        if (!this.player.isPlaying())
-            return;
+        try
+        {
+            JSONObject timeObj=new JSONObject();
+            timeObj.put("current", curSecs);
+            timeObj.put("total", totalSecs);
+            JSONObject msgObj=new JSONObject();
+            msgObj.put("type", "time");
+            msgObj.put("time", timeObj);
 
-        //sendMessage({type:"time", time:{current:c, total:t}});
-        double curSecs = (double)(this.player.getCurrentTimeMillis()/1000);
-        double totalSecs = (double)(this.player.getDurationMillis()/1000);
-        JSONObject timeObj = new JSONObject();
-        timeObj.put("current",curSecs);
-        timeObj.put("total", totalSecs);
-        JSONObject msgObj = new JSONObject();
-        msgObj.put("type", "time");
-        msgObj.put("time", timeObj);
-
-        sendMessage(msgObj);
+            sendMessage(msgObj);
+        }
+        catch(Throwable t)
+        {
+            Log.w(Extension.TAG, "Could not send time data", t);
+        }
     }
 
     /** Send player state to Actionscript */
-    private void sendState(String state)
+    public void sendState(String state)
     {
         //sendMessage({type:"state", state:state});
         try
@@ -713,7 +245,7 @@ public class AndroidYoutubeContext extends FREContext implements YouTubePlayer.O
     }
 
     /** Send Error to Actionscript */
-    private void sendError(int code, String msg)
+    public void sendError(int code, String msg)
     {
         //sendMessage({type:"err", code:e.data, msg:msg})
         try
@@ -737,12 +269,6 @@ public class AndroidYoutubeContext extends FREContext implements YouTubePlayer.O
      */
     public void dispatchEventWithReason(String type, String reason)
     {
-        if (!checkState(State.READY))
-        {
-            Log.w(Extension.TAG, "Sending event to AIR App when not ready ("+type+", "+reason+")");
-            return;
-        }
-
         dispatchStatusEventAsync(type, reason);
     }
 
@@ -776,106 +302,6 @@ public class AndroidYoutubeContext extends FREContext implements YouTubePlayer.O
             this.unsupportedReason = "YouTube App not installed";
             Log.w(Extension.TAG, "YouTube App not installed.");
             return false;
-        }
-    }
-
-    /** Atomic Integer for generating View ID's */
-    private static final AtomicInteger sNextGeneratedId = new AtomicInteger(1);
-
-    /** Generate a View ID with no conflicts */
-    public static int generateViewId()
-    {
-        //HACK this is taken from View.generateViewId() but that is API 17+
-        for (;;)
-        {
-            final int result = sNextGeneratedId.get();
-            // aapt-generated IDs have the high byte nonzero; clamp to the range under that.
-            int newValue = result + 1;
-            if (newValue > 0x00FFFFFF) newValue = 1; // Roll over to 1, not 0.
-            if (sNextGeneratedId.compareAndSet(result, newValue))
-            {
-                return result;
-            }
-        }
-    }
-
-    // Nested Classes
-
-    /** Video Timer */
-    class VideoTimer
-    {
-        /** Update time */
-        private long updateTime;
-
-        /** Runnable */
-        private Runnable runnable;
-
-        /** Handler */
-        private Handler handler;
-
-        /** Create a new VideoTimer */
-        public VideoTimer(final long updateTime)
-        {
-            // Handler created/run on UI Thread
-            this.updateTime = updateTime;
-            this.handler = new Handler();
-            this.runnable = new Runnable()
-            {
-                @Override
-                public void run()
-                {
-                    try
-                    {
-                        sendTimeData();
-                    }
-                    catch (Throwable t)
-                    {
-                        Log.e(Extension.TAG, "Could not send time data", t);
-                    }
-                    handler.postDelayed(runnable, updateTime);
-                }
-            };
-        }
-
-        /** Start */
-        public void start()
-        {
-            // ensure not already started to prevent doubling up callbacks
-            stop();
-            handler.postDelayed(runnable, updateTime);
-        }
-
-        /** Stop */
-        public void stop()
-        {
-            handler.removeCallbacks(runnable);
-        }
-    }
-
-    /** YouTube Player Fragment */
-    public static class CustomPlayerFragment extends YouTubePlayerFragment
-    {
-        /** On Create View */
-        @Override public View onCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState)
-        {
-            Log.d(Extension.TAG, "CustomPlayerFragment.onCreateView()");
-
-            Extension.context.onFragmentViewCreated();
-
-            return super.onCreateView(inflater, container, savedInstanceState);
-        }
-
-        /** On Save Instance State */
-        @Override public void onSaveInstanceState(Bundle bundle)
-        {
-            Log.d(Extension.TAG, "CustomPlayerFragment.onSaveInstanceState()");
-
-            //HACK: This is in place to fix missing parcelable/no class def found errors.
-            //see: https://stackoverflow.com/questions/44558166/fatal-exception-java-lang-noclassdeffounderror-rt
-            //and: https://stackoverflow.com/questions/44379747/youtube-android-player-api-throws-badparcelableexception-classnotfoundexception
-            Extension.context.onFragmentSaveInstance(false);
-
-            super.onSaveInstanceState(bundle);
         }
     }
 
@@ -931,7 +357,7 @@ public class AndroidYoutubeContext extends FREContext implements YouTubePlayer.O
         protected int execute(FREObject[] params) throws Exception
         {
             String videoId = params[0].getAsString();
-            playVideoById(videoId, 0.0f);
+            adapter.execute(createPlayByIdAction(videoId,0.0f));
             return 0;
         }
     }
@@ -972,7 +398,7 @@ public class AndroidYoutubeContext extends FREContext implements YouTubePlayer.O
             int frameY = params[1].getAsInt();
             int frameWidth = params[2].getAsInt();
             int frameHeight = params[3].getAsInt();
-            setVideoFrame(frameX, frameY, frameWidth, frameHeight);
+            adapter.setVideoFrame(frameX, frameY, frameWidth, frameHeight);
             return 0;
         }
     }
@@ -991,7 +417,11 @@ public class AndroidYoutubeContext extends FREContext implements YouTubePlayer.O
         protected int execute(FREObject[] params) throws Exception
         {
             String devKey = params[0].getAsString();
-            initVideo(devKey);
+
+            adapter=YouTubePlayerAdapter.fetchInstance();
+
+            adapter.init(getActivity(),devKey,AndroidYoutubeContext.this);
+
             return 0;
         }
     }
